Fix some compiler warnings.

+++ src/net/common/servergamestate.cpp
@@ -1357,7 +1357,7 @@ ServerGameStateWaitPlayerAction::TimerTimeout(const boost::system::error_code &e
 			curPlayer->incrementActionTimeoutCounter();
 
 			if (server->GetGameData().gameType == GAME_TYPE_RANKING
-				&& server->GetGameData().playerActionTimeoutSec >= server->GetSmallDelaySec())
+				&& server->GetGameData().playerActionTimeoutSec >= (int)server->GetSmallDelaySec())
 			{
 				if (curPlayer->getActionTimeoutCounter() == SERVER_WARNING_ACTION_TIMEOUT_THRESHOLD)
 				{
